Let higher versions of pandas as a dependency.
I'm not aware of any reason why higher versions of Pandas cannot be
used as a dependency, so I'm letting higher versions.

The current pinned version we have is causing me trouble with
Python 3.7, apparently the combination of dependencies we have
causes trouble for scipy (which is a dependency of Pandas)
to be properly installed.